Committing Java UML Parser
‘JavaUMLParser’ is developed using Doclets that uses the doclet API to
parse source Java files and retrieve their Classes, Methods, Attributes
and Constructors, along with additional details such as associations and
dependencies.

We use these details to generate the output UML diagram using Plantuml
which is an Open Source project that allows us to generate Class diagram
based on a predefined format input. This requires the Graphviz software
to be installed to be able to generate an image and save it in an Output
file.